// src/api/dtos/productForAdminDTO.ts
/**
 * Generated by orval v7.0.1 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { MoneyDTO } from './moneyDTO';

export interface ProductForAdminDTO {
  createdDate?: string;
  id?: string;
  lastModifiedDate?: string;
  name?: string;
  previewImageURL?: string;
  unitPrice?: MoneyDTO;
}